Keep track of shadow APU registers and a linear counter
This improves accuracy for note cutoffs for the triangle channel,
by calculating a decent approximation for the linear counter's
current position at draw time. Mesen's Lua API does not provide
this value, and the length counter that it does provide is in
most cases 2x too slow, which was previously causing triangle
notes to visibly cut off long after the channel was audibly
silenced by the hardware.

Additionally, this corrects a small bug with the noise channel,
and provides a debug feature to display the APU registers, though
I find this feature ugly and want to touch it up before the
next release.