Apply icon highlighting functionality to envelope and sweep indicators
Also, fix a bug with enable detection in the sweep unit; if the shift
amount is 0, even if the enabled flag is set, the sweep unit has no
effect other than to apply muting logic. This isn't terribly interesting
musically, so I've chosen not to highlight the indicator in this case.